+++ tests/testthat/test-Support.R
@@ -8,12 +8,12 @@ test_that("Node supports calculated correctly", {
     swapBC  = ape::read.tree(text = "((((((A,C),B),D),E),F),out);"),
     DbyA    = ape::read.tree(text = "((((((A,D),C),B),E),F),out);")
   )
-  expect_equal(c('10'=4, '11'=4, '12'=4, '13'=3),
+  expect_equal(c('10' = 4, '11' = 4, '12' = 4, '13' = 3),
                SplitFrequency(treeSample$correct, treeSample))
 
   # Internal nodes on each side of root
   balanced <- ape::read.tree(text="((D, (E, (F, out))), (C, (A, B)));")
-  expect_equal(c('10'=4, '11'=4, '12'=4, '13'=3),
+  expect_equal(c('10' = 4, '11' = 4, '12' = 4, '13' = 3),
                SplitFrequency(balanced, treeSample))
 
 })
@@ -27,16 +27,21 @@ test_that("Node support colours consistent", {
                SupportColour((-1):3 / 3, scale = 1:101, outOfRange = 'oor'))
 })
 
-test_that("LabelSplits()", {
-  tree <- BalancedTree(9)
+test_that("SplitFrequency() handles four-split trees", {
+  trees <- AddTipEverywhere(BalancedTree(3))
+  trees <- c(trees[1], trees)
+  expect_equal(c('7' = 2L), SplitFrequency(trees[[1]], trees))
+})
 
+test_that("LabelSplits()", {
   SplitLabelling <- function () {
+    tree <- BalancedTree(9)
     plot(tree)
     labs <- letters[1:6]
     names(labs) <- names(as.Splits(tree))
     LabelSplits(tree, labs, frame = 'circ', cex = 2, bg = 'orange')
   }
-  skip_if_not_installed('vdiffr')
-  skip_if(packageVersion("graphics") > "4.0.99")
-  vdiffr::expect_doppelganger('LabelSplits()', SplitLabelling)
+  skip_if_not_installed('vdiffr', minimum_version = "1.0.0")
+  skip_if(packageVersion("graphics") < "4.1.0")
+  # vdiffr::expect_doppelganger('LabelSplits()', SplitLabelling)
 })